+++ src/js/components/Profile.jsx
@@ -1,10 +1,8 @@
 // @flow
 
 import React from 'react';
-import { connect } from 'react-redux';
 import LoadingDots from './LoadingDots';
 import CurrentUser from './CurrentUser';
-import { storyValue, createStory, clearInput } from '../actions/profile';
 
 const Profile = (props: {
   currentUser: Object,
@@ -14,7 +12,6 @@ const Profile = (props: {
   handleSubmit: Function
 }) => {
   const { currentUser, value, profile, handleChange, handleSubmit } = props;
-  console.log(profile.story, ' == ', value);
   return (
     <div>
       {!currentUser
@@ -30,22 +27,4 @@ const Profile = (props: {
   );
 };
 
-const mapStateToProps = ({ authentication, currentUser, value, profile }) => ({
-  authentication,
-  currentUser,
-  value,
-  profile
-});
-
-const mapDispatchToProps = (dispatch: Function) => ({
-  handleChange(e) {
-    dispatch(storyValue(e.target.value));
-  },
-  handleSubmit(e, story, uid) {
-    e.preventDefault();
-    dispatch(createStory({ story, uid }));
-    dispatch(clearInput());
-  }
-});
-
-export default connect(mapStateToProps, mapDispatchToProps)(Profile);
+export default Profile;
