+++ src/js/containers/MeContainer.jsx
@@ -1,0 +1,8 @@
+// @flow
+
+import { connect } from 'react-redux';
+import Me from '../components/Me';
+
+const mapStateToProps = ({ currentUser, profile }) => ({ currentUser, profile });
+
+export default connect(mapStateToProps)(Me);
